Fix all exotic characters in IDs
I have a source document with backslashes in the file name (windows paths). But there are also lots of other characters that would be invalid. I propose replacing all non-alphanumerics with underscores.
The full ID rules are quite complex, so this feels like a good middle ground between the current hardcoded replacements and a "proper" ruleset.